% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_helpers.R
\name{cheap_darken}
\alias{cheap_darken}
\alias{darken_face}
\title{'light' effect helper functions}
\usage{
cheap_darken(col, amount)

darken_face(
  face,
  col,
  top = 0,
  west = 0.2,
  east = 0.2,
  south = 0.4,
  north = 0.4,
  darken_fn = cheap_darken
)
}
\arguments{
\item{col}{Color to darken}

\item{amount}{Fraction to darken by}

\item{face}{Cube/cuboid face to color.  One of "top", "west", "east", "south", or "north".}

\item{top}{Amount to darken the "top" face.}

\item{west}{Amount to darken the "west" face.}

\item{east}{Amount to darken the "east" face.}

\item{south}{Amount to darken the "south" face.}

\item{north}{Amount to darken the "north" face.}

\item{darken_fn}{Function to darken with.  Should take two arguments:
the first should be the colour and the second should be numeric amount to darken by.
Default will be to use \code{\link[=cheap_darken]{cheap_darken()}}.
\code{colorspace::darken()} is a slower, \dQuote{better} alternative.}
}
\value{
Vector of darkened colors.
}
\description{
Helper functions to generate a \dQuote{light} effect
for \code{\link[=oblicubesGrob]{oblicubesGrob()}}, \code{\link[=grid.oblicubes]{grid.oblicubes()}}, \code{\link[=oblicuboidsGrob]{oblicuboidsGrob()}}, and \code{\link[=grid.oblicuboids]{grid.oblicuboids()}}.
\code{darken_face()} is the default \code{light} argument for \code{\link[=oblicubesGrob]{oblicubesGrob()}}, \code{\link[=grid.oblicubes]{grid.oblicubes()}},
\code{\link[=oblicuboidsGrob]{oblicuboidsGrob()}}, and \code{\link[=grid.oblicuboids]{grid.oblicuboids()}}.
\code{cheap_darken()} is the default darkening function used by \code{darken_face()}.
}
\details{
The \code{light} argument of \code{\link[=oblicubesGrob]{oblicubesGrob()}}, \code{\link[=grid.oblicubes]{grid.oblicubes()}}, \code{\link[=oblicuboidsGrob]{oblicuboidsGrob()}}, and \code{\link[=grid.oblicuboids]{grid.oblicuboids()}}
needs a function that that takes two arguments: the first is \code{face} one of its five faces:
"top", "west", "east", "south", or "north" and the second is \code{col} the cube/cuboid's fill color
}
\examples{
  cheap_darken("blue", 0.5)
  darken_face("top", "blue")
}
