% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicubes-package.R
\docType{package}
\name{oblicubes-package}
\alias{oblicubes}
\alias{oblicubes-package}
\title{oblicubes: Obliquely Projected Cubes and Cuboids}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for 3D rendering in 'grid' and 'ggplot2' graphics via an oblique projection using cubes and cuboids. Also supports 2D rendering via a primary view orthographic projection. Can be viewed as an extension to the 'isocubes' package \url{https://github.com/coolbutuseless/isocubes}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://trevorldavis.com/R/oblicubes/}
  \item Report bugs at \url{https://github.com/trevorld/oblicubes/issues}
}

}
\author{
\strong{Maintainer}: Trevor L Davis \email{trevor.l.davis@gmail.com} (\href{https://orcid.org/0000-0001-6341-4639}{ORCID})

Authors:
\itemize{
  \item Mike FC (Some code adapted from isocubes)
}

}
\keyword{internal}
